+++ CHANGELOG.rst
@@ -25,6 +25,8 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 Unreleased
 ------------------------------------------------------------------------------
 
+* |Enhancement| Added ``chunksize`` parameter for piecemeal data reading to builtin DataSources,
+  `issue #120 <https://github.com/schuderer/mllaunchpad/issues/120>`_,
 * |Fixed| ``mllaunchpad --verbose`` now correctly logs DEBUG information,
   `issue #119 <https://github.com/schuderer/mllaunchpad/issues/119>`_,
 * |Fixed| Fixed an issue where builtin DataSources could not be found when configured,